requirements.txt
streamlit>=1.28.0
pillow>=9.0.0
torch>=2.0.0
transformers>=4.30.0
huggingface_hub>=0.17.0
opencv-python>=4.7.0
python-dotenv>=1.0.0
numpy>=1.24.0
matplotlib>=3.7.0
requests>=2.28.0
tqdm>=4.65.0
# Added for speech recognition using Hugging Face Transformers
librosa>=0.10.0
accelerate>=0.20.0
# Added for asyncio fix
nest_asyncio>=1.5.6
# Make sure we have the correct imageio-ffmpeg package
imageio-ffmpeg>=0.4.8
# Add new dependencies for audio extraction
moviepy==1.0.3
soundfile>=0.10.3 